Created border color control group and added the border color matches edge color items to it.
